Massively buffed  lux absorption of reinforced glass.
A lux-absorb value of 20 provides 20/64 absorption, or just over 31%. This value is likely much lower than WintersKnight intended, and is in fact less than half as effective as the 83% absorption it would have due to its cracky value if the line was omitted. I've given it the metallic group (100% absorption) as that seems to be the intention.